Рефакторинг UI-компонентов, фикс перетаскивания GraphML на окно (#156)
