+++ src/renderer/src/components/MainContainer/Tabs/Tabs.tsx
