+++ src/renderer/src/components/MainContainer/Tabs/index.ts
