+++ src/renderer/src/components/MainContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './MainContainer';
