+++ src/renderer/src/components/SaveRemindModal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useForm } from 'react-hook-form';
 
-import { Modal } from './Modal/Modal';
+import { Modal } from '@renderer/components/UI';
 
 interface SaveRemindModalProps {
   isOpen: boolean;
