+++ src/renderer/src/components/ScrollableList.tsx
@@ -1,4 +1,5 @@
 import React, { ComponentProps, useEffect, useRef, useState } from 'react';
+
 import { twMerge } from 'tailwind-merge';
 
 interface ScrollableListProps<T> {
