+++ src/renderer/src/components/UI/Tabs/index.ts
@@ -1,0 +1,2 @@
+export * from './Tabs';
+export * from './TabPanel';
