+++ src/renderer/src/components/UI/index.ts
@@ -1,2 +1,7 @@
 export * from './Badge';
 export * from './Select';
+export * from './Tabs';
+export * from './WithHint';
+export * from './Modal';
+export * from './ColorInput';
+export * from './TextInput';
