+++ src/renderer/src/components/index.ts
@@ -1,4 +1,3 @@
-export * from './Modal';
 export * from './CodeEditor';
 export * from './CreateModal';
 export * from './DiagramEditor';
@@ -8,8 +7,8 @@ export * from './CreateSchemeModal';
 export * from './serverSelect/FlasherSelectModal';
 export * from './ErrorModal';
 export * from './Sidebar';
-export * from './Tabs';
 export * from './MainContainer';
 export * from './ComponentAddModal';
 export * from './ComponentEditModal';
 export * from './ComponentDeleteModal';
+export * from './Scale';
