+++ src/renderer/src/components/serverSelect/DocSelectModal.tsx
@@ -2,8 +2,7 @@
 import { useForm } from 'react-hook-form';
 import { twMerge } from 'tailwind-merge';
 
-import { Modal } from '../Modal/Modal';
-import { TextInput } from '../Modal/TextInput';
+import { Modal, TextInput } from '@renderer/components/UI';
 import { Settings } from '../Modules/Settings';
 
 interface DocSelectModalProps {
