+++ src/renderer/src/components/serverSelect/FlasherSelectModal.tsx
@@ -2,9 +2,8 @@ import { Controller, useForm } from 'react-hook-form';
 
 import { Select } from '@renderer/components/UI';
 
-import { Modal } from '../Modal/Modal';
-import { TextInput } from '../Modal/TextInput';
 import { Settings } from '../Modules/Settings';
+import { Modal, TextInput } from '@renderer/components/UI';
 
 const SELECT_LOCAL = 'local';
 const SELECT_REMOTE = 'remote';
